requirements.txt
aiogram==3.5.0
requests==2.31.0
loguru==0.7.2
apscheduler==3.10.4
python-dotenv==1.0.1
peewee==3.17.3
fake-useragent==1.5.1
beautifulsoup4==4.12.3
lxml==5.2.1